server/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.2
flask-cors==6.0.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
Werkzeug==3.1.3

# Database & ORM
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10

# Migrations
Flask-Migrate==4.1.0

# Environment variables
python-dotenv==1.0.1

faker==37.8.0 